moved validateNodes function to utils
